+++ app/Http/Admin/Services/Trade.php
@@ -5,6 +5,7 @@ namespace App\Http\Admin\Services;
 use App\Builders\TradeList as TradeListBuilder;
 use App\Library\Paginator\Query as PaginateQuery;
 use App\Models\Refund as RefundModel;
+use App\Models\Trade as TradeModel;
 use App\Repos\Account as AccountRepo;
 use App\Repos\Order as OrderRepo;
 use App\Repos\Trade as TradeRepo;
@@ -13,6 +14,16 @@ use App\Validators\Trade as TradeValidator;
 
 class Trade extends Service
 {
+
+    public function getChannelTypes()
+    {
+        return TradeModel::channelTypes();
+    }
+
+    public function getStatusTypes()
+    {
+        return TradeModel::statusTypes();
+    }
 
     public function getTrades()
     {